refactor(type): adjusted type names
