--- src/types/Programs.ts
@@ -1,5 +1,0 @@
-export enum Program {
-  CS,
-  IT,
-  IS,
-}
