+++ src/types/ProgramsEnum.ts
@@ -1,0 +1,5 @@
+export enum ProgramEnum {
+  CS = "CS",
+  IT = "IT",
+  IS = "IS",
+}
